Split and migrate build scripts to ES6
#1
