+++ .gitignore
@@ -1,5 +1,6 @@
 # build artifacts
 dist/
+cli.js
 
 # node modules
 node_modules/
