+++ server/cli.js
@@ -1,19 +1,20 @@
 /**
  * CLI runner for together system
  */
-var program = require('commander');
-var _ = require('lodash');
+import program from 'commander';
+import startApp from './app.js';
+import defaultConfig from '../config.js';
 
 program
   .version('0.1.0')
   .description('Starts server and show photos from specified dir')
   .arguments('<photos-dir>')
   .option('-p --port [port]', 'Server port', 8000)
-  .option('-w --cwd <dir>', 'Application working directory', __dirname);
+  .option('-w --cwd <dir>', 'Application working directory', './');
 
 program.parse(process.argv);
 
-if (program.cwd !== process.cwd) {
+if (program.cwd !== process.cwd()) {
   process.chdir(program.cwd);
 }
 
@@ -26,12 +27,11 @@ if (!photosDir) {
   process.exit(1);
 }
 
-var config = _.extend({}, require('../config'));
+var config = Object.assign({}, defaultConfig);
 config.development = false;
 config.port = port;
 config.photosDir = photosDir;
 config.previewLoader.type = 'filesystem';
 config.previewLoader.config.photosDir = photosDir;
 
-var startApp = require('./app');
 startApp(config);
